Consolidate remote config tests into same directory/file
