+++ tests/remote_config/test_remote_configuration.py
@@ -21,6 +21,25 @@ from utils import (
 from utils.tools import logger
 
 
+@rfc("https://docs.google.com/document/d/1bUVtEpXNTkIGvLxzkNYCxQzP2X9EK9HMBLHWXr_5KLM/edit#heading=h.vy1jegxy7cuc")
+@features.remote_config_object_supported
+class Test_NoError:
+    """A library should apply with no error all remote config payload."""
+
+    def test_no_error(self):
+        def no_error(data):
+            config_states = (
+                data.get("request", {}).get("content", {}).get("client", {}).get("state", {}).get("config_states", {})
+            )
+
+            for state in config_states:
+                error = state.get("apply_error", None)
+                if error is not None:
+                    raise Exception(f"Error in remote config application: {error}")
+
+        interfaces.library.validate_remote_configuration(no_error, success_by_default=True)
+
+
 @rfc("https://docs.google.com/document/d/1u_G7TOr8wJX0dOM_zUDKuRJgxoJU_hVTd5SeaMucQUs/edit#heading=h.octuyiil30ph")
 @features.remote_config_object_supported
 class RemoteConfigurationFieldsBasicTests:
